Move address validation to OutgoingSmsReceiver
It turns out that by the time the validation got to VoicePlusService, we
had already canceled the intent, making it impossible to then pass it
onward to cellular (the intended behavior). The obvious solution is to
check to see whether it's a GVoice-able address *before* we cancel the
intent, and only cancel it if it is.

In other words, this patch makes review #50850 behave as intended.

Change-Id: I8b30605f89600814c5a1a223bd562c17f003d373